extract TestIO into spec folder
